requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.6.0
    # via pydantic
attrs==23.2.0
    # via
    #   cattrs
    #   lsprotocol
cattrs==23.2.3
    # via
    #   lsprotocol
    #   pygls
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
lsprotocol==2023.0.1
    # via pygls
pydantic==2.7.1
pydantic-core==2.18.2
    # via pydantic
pygls==1.3.1
result==0.16.1
smmap==5.0.1
    # via gitdb
typing-extensions==4.11.0
    # via
    #   pydantic
    #   pydantic-core
